+++ src/irdl/__init__.py
@@ -1,6 +1,4 @@
 """Impulse Response Downloader (irdl): Download, unpack and process impulse response datasets."""
 
-from .logger import logger
-
 from .ista import MiracleDataset, SrirachaDataset
 from .sofa import FabianDataset
