+++ src/irdl/cli.py
@@ -10,7 +10,7 @@ from numpydoc.docscrape import FunctionDoc
 
 import irdl
 from irdl.base import BaseDataset
-from irdl.logger import configure_cli_logging, logger
+from irdl.logging import configure_cli_logging
 
 # Configure CLI logging
 configure_cli_logging()
@@ -55,7 +55,7 @@ def _get_dataset_classes():
     return dataset_classes
 
 
-# Typer app that can be invoked by calling ``irdl`` from the CLI.
+#: Typer app that can be invoked by calling ``irdl`` from the CLI.
 app = typer.Typer(no_args_is_help=True)
 
 # Automatically register all supported datasets as subcommands to the app.
@@ -76,7 +76,7 @@ for dataset_class in _get_dataset_classes():
     def make_wrapper(cls, method, params, help_text, dataset_name, param_docs):
         def wrapper(**kwargs):
             return method.__func__(cls, **kwargs)
-        
+
         # Build the signature for the wrapper
         new_params = []
         for name, p in params.items():
@@ -85,16 +85,14 @@ for dataset_class in _get_dataset_classes():
             resolved_type = _resolve_union_type(p.annotation)
             # Don't pass default to typer.Option - it's already in the parameter
             new_params.append(
-                p.replace(
-                    annotation=Annotated[resolved_type, typer.Option(help=param_docs.get(name, ""))]
-                )
+                p.replace(annotation=Annotated[resolved_type, typer.Option(help=param_docs.get(name, ""))])
             )
-        
+
         wrapper.__signature__ = signature(wrapper).replace(parameters=new_params)
         wrapper.__doc__ = help_text
         wrapper.__name__ = f"{dataset_name}_wrapper"
         return wrapper
-    
+
     wrapper = make_wrapper(dataset_class, get_method, sig.parameters, help_text, dataset_class.name, param_docs)
 
     # Register subcommand using dataset_class.name for the command name
